+++ source/forms/frmPal.vb
@@ -1,72 +1,82 @@
-﻿Public Class frmPal
+﻿''' <summary>
+''' Color Palette window
+''' </summary>
+Public Class FrmPal
 
-    Private Sub ToolStripStatusLabel1_Click(sender As Object, e As EventArgs) Handles ssFileName.Click
+    ''' <summary>
+    ''' Handles color replacement in main color palette (in main window). 
+    ''' This allows to quickly preview how graphics would look like too, with a different color pattern.
+    ''' </summary>
+    ''' <param name="sender">Object</param>
+    ''' <param name="e">EventArgs</param>
+    Private Sub BtnUseInMainPal_Click(sender As Object, e As EventArgs) Handles BtnUseInMainPal.Click
 
-    End Sub
+        ' If no main graphic has been loaded or created with at least 1 frame, there is nothing to replace.
+        If IsNothing(EditorGraphic.Frames) Then
 
-
-    Private Sub BtnUseInMainPal_Click(sender As Object, e As EventArgs) Handles btnUseInMainPal.Click
-
-
-        'g2.colorPalette.colors().Length
-        If IsNothing(editorGraphic.Frames) Then
-
-            MsgBox("You will need to open a ZT1 graphic file first," & vbCrLf & "then you can use the recolor feature.",
-                   vbOKOnly + vbCritical,
-                   "No graphic available")
-
+            MdlZTStudio.ExpectedError(Me.GetType().FullName, "BtnUseInMainPal_Click", "You will need to open a ZT1 graphic file first before you can use this recolor feature.")
             Exit Sub
 
         End If
 
+        Dim IntMaxColorIndex As Integer = FrmMain.DgvPaletteMain.Rows.Count - Me.DgvPal.Rows.Count + 1
 
-        Dim Input As String = InputBox("" &
-                "Index of first color to replace (can not be 0 since we ignore transparent colors)." & vbCrLf &
-                "For example, the index for the Restaurant would be 248 (roof 8 colors) or 232 (roof 16 colors)", "Index of the first color to replace", "1")
-        If Input = vbNullString Then Exit Sub 'user chickened out
+        ' Can't be negative (main color palette color count is smaller than the count of this palette)
+        If IntMaxColorIndex < 1 Then
+            MdlZTStudio.ExpectedError(Me.GetType().FullName, "BtnUseInMainPal_Click", "The color palette in the main window has less colors than this palette. Replacing is not possible.")
+            Exit Sub
+        End If
 
-        Dim IntStart As Integer = CInt(Input)
+        Dim StrMessage As String = "" &
+                "Index of first color to replace (can not be 0 since the transparent color is ignored)." & vbCrLf & vbCrLf &
+                "For example, the index for the Restaurant would be 248 (roof 8 colors) or 232 (other roof 16 colors)" & vbCrLf & vbCrLf &
+                "With the current palette, the index to start from should be between 1 and " & IntMaxColorIndex
 
-        'If intStart > (editorGraphic.colorPalette.colors.Count - dgvPal.Rows.Count + 1) Then
-        'MsgBox("The index is too high. The max index for this image is " & (editorGraphic.colorPalette.colors.Count - dgvPal.Rows.Count + 1))
-        'Exit Sub
-        'End If
+        Dim strInput As String = InputBox(StrMessage, "Index of the first color to replace", "1")
 
-        If (intStart + dgvPal.Rows.Count - 1) > 256 Then
-            MsgBox("ZT1 Color palettes are limited to 256 colors.", vbOKOnly + vbCritical, "Can't replace colors")
+        If strInput = vbNullString Then
+            Exit Sub 'user pressed cancel
+
+        ElseIf IsNumeric(strInput) = False Then
+
+            ' Verify
+            MdlZTStudio.ExpectedError(Me.GetType().FullName, "BtnUseInMainPal_Click", "You need to specify a positive number, at least 1 and maximum " & intmaxcolorindex & ".")
+            Exit Sub
+
+        ElseIf Int(strInput) > intmaxcolorindex Then
+
+            MdlZTStudio.ExpectedError(Me.GetType().FullName, "BtnUseInMainPal_Click", "You need to specify a positive number, at least 1 and maximum " & intmaxcolorindex & ".")
             Exit Sub
 
         End If
-
-
 
         ' Replace
-        For Each drRow As DataGridViewRow In dgvPal.Rows
-            ' Ignore first = transparent color. It doesn't matter.
-            ' However, when we are replacing a color, it DOES matter.
-            ' Pal8 and Pal16 files contain the transparent color, then the 8 or 16 shades of a color.
+        For Each ObjDataRow As DataGridViewRow In DgvPal.Rows
 
+            ' Careful: the game uses color palettes consisting of 1 transparent color; followed by 8 or 16 shades of a color for color customization.
 
-            If drRow.Index <> 0 Then ' dummy color
+            If ObjDataRow.Index <> 0 Then ' Transparent color, ignore this, it doesn't matter.
 
-                If editorGraphic.ColorPalette.Colors.Count = (intStart + drRow.Index - 1) Then
+                If EditorGraphic.ColorPalette.Colors.Count = (CInt(strInput) + ObjDataRow.Index - 1) Then
                     ' Color index does not exist.
-                    editorGraphic.ColorPalette.Colors.Add(drRow.DefaultCellStyle.BackColor)
+                    EditorGraphic.ColorPalette.Colors.Add(ObjDataRow.DefaultCellStyle.BackColor)
                 Else
                     ' Color index already existed. Overwrite.
-                    editorGraphic.ColorPalette.Colors(intStart + drRow.Index - 1) = drRow.DefaultCellStyle.BackColor
+                    EditorGraphic.ColorPalette.Colors(CInt(strInput) + ObjDataRow.Index - 1) = ObjDataRow.DefaultCellStyle.BackColor
                 End If
 
             End If
         Next
 
-        ' Force redraw. In this case, we should ignore the rendered images and make them regenerate from HEX.
-        For Each ztFrame As ClsFrame In editorGraphic.Frames
-            ztFrame.CoreImageBitmap = Nothing
+        ' The color palette in the main window has been changed.
+        ' This means cached versions of images (CoreImageBitmap) should be regenerated from hex values.
+        ' SO, set CoreImageBitMap to Nothing (will be rendered when needed; not instantly!)
+        For Each ObjFrame As ClsFrame In EditorGraphic.Frames
+            ObjFrame.CoreImageBitmap = Nothing
         Next
 
 
-        editorGraphic.LastUpdated = Now.ToString("yyyyMMddHHmmss")
+        EditorGraphic.LastUpdated = Now.ToString("yyyyMMddHHmmss")
         MdlZTStudioUI.UpdatePreview(True, False, FrmMain.TbFrames.Value - 1)
 
         ' Show palette again?
@@ -75,13 +85,5 @@
 
     End Sub
 
-    Private Sub DgvPal_CellContentClick(sender As Object, e As DataGridViewCellEventArgs) Handles dgvPal.CellContentClick
 
-    End Sub
-
-    Private Sub MnuExport_GIMP_TextFile_Click(sender As Object, e As EventArgs)
-
-
-
-    End Sub
 End Class